Heap uses a more efficient tree
The heap now stores the original index per item in a vector, rather than in the tree itself. This should allow significantly more efficient traversal of the tree. This version also separates the heap selector from indexed_collection.hpp, because this library doesn't need a general-purpose indexed collection - it's only necessary for the heap selector, and so it can be spun off and combined into the heap selector in its entirety.