Yeah, okay, disregard previous commits
Total weight wasn't being calculated, and so the selector always returned index 1. With that fixed, the static version outperforms std::discrete_distribution, but only until discrete_distribution starts to slow down - then the Fenwick tree gets even worse than it. Changing distributions are consistently predictable, at least, but are significantly worse than on the traditional tree structure. This is an interesting datastructure, but it's not exactly promising. Alas.