Segment tree
Implemented a segment tree in place of the tree we were using before. Sorry for not making new classes for it! Currently it just replaces the entire tree structure with itself. Tests have been updated to not do the heap, because heaping is impossible in this version. Heap files also remain in the commit, they're just not used. Performance-wise, this thing is weird - it seems to get "good" performance when static, nearly perfectly on par with the standard library (until larger datasets, when it gets better and better comparably), but then it seems to be unbelievably close to O(1) when the data changes as it's selected - needs much more interrogation. Prints are commented out in weightsum_tree.hpp; if you want to verify the logic as it goes through and selects and updates, feel free to just turn those back on.