Updated style to match
Complete tree has a new function that needs to be kept, but otherwise, everything is in single modular files. Leafsum tree should be about equivalent to the previous implementation (hopefully), and there's also now leafsum tree split, which separates the leaves and the rest of the tree for better cache locality.