Added pushing and popping
Both leaf_sum_tree and leaf_sum_tree_split now support pushing and popping, through push_back() and pop_back(). They also both have overloads for mass pushing and popping now. Both files do their best to be efficient, but unfortunately expanding or shrinking the datastructure size itself takes a lot of work. It's easier in some ways on the split version, since you don't have to deal with grabbing the leaves and then splicing them back in. Another optimization to the split version - leaves is now only the size of the number of weights in it, and thus is no longer always the size of the next power of two. This required a safety check, but should be a little more memory efficient. Core logic of both remains unchanged.