Should be fully std:: compatible now!
Implemented all of the std functions to my knowledge right now. For the sake of cleanness, the assert calls currently in the code should be taken out or made throws instead, but that's not important yet. This commit also includes two new testing files, for testing more or less standard operation and accuracy of the library. Goes toe-to-toe with std::discrete_distribution and returns exactly the same outputs! :]