Fixed resizing
"Exponential tree" is actually a k-ary tree, and I fixed its resizing and assignment. It now resizes only to the size needed for however many weights it has, instead of whatever the nearest power of fanout is. I very much need to go through and change all the terms and variable names, cause they should not be referencing exponential concepts.